change idbig category table

+++ database/migrations/2019_05_21_102411_add_field_id_category_id.php
@@ -15,7 +15,7 @@ class AddFieldIdCategoryId extends Migration
     {
         Schema::table('products', function (Blueprint $table) {
             //
-            $table->unsignedInteger('id_category');
+            $table->unsignedBigInteger('id_category');
             $table->foreign('id_category')
                 ->references('id')
                 ->on('categories')
